requirements.txt
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.21.0
peft>=0.4.0
transformers>=4.30.0
accelerate>=0.20.0
datasets>=2.12.0
wandb>=0.15.0
opencv-python>=4.8.0
Pillow>=9.5.0
numpy>=1.24.0
scikit-learn>=1.3.0
svgpathtools>=1.6.0
svgwrite>=1.4.0
tqdm>=4.65.0
hydra-core>=1.3.0
omegaconf>=2.3.0
rich>=13.0.0
huggingface-hub>=0.16.0
safetensors>=0.3.0
xformers>=0.0.20
bitsandbytes>=0.41.0